Solga

Serving APIs

serve :: Router r => r -> Application

Serve a Router with Solga, returning SolgaErrors as HTTP responses.

serveThrow :: Router r => r -> Application

Serve a Router with Solga, throwing SolgaErrors.

Basic routers

type (:>) f g = f g infixr 2

Compose routers. This is just type application, ie.: Foo :> Bar :> Baz == Foo (Bar Baz)

type (/>) seg g = Seg seg :> g infixr 2

Match a path, segment, e.g "foo" /> JSON Bar

type Get a = Endpoint "GET" (JSON a)

Handle a GET request and produce a JSON response, with IO.

type Post a = Endpoint "POST" (JSON a)

Handle a POST request and produce a JSON response, with IO.

newtype JSON a

Return a given JSON object

Constructors

JSON

Fields

jsonResponse :: a

newtype Raw

Serve a given WAI Application.

Constructors

Raw

Fields

rawApp :: Application

newtype RawResponse

Serve a given WAI Response.

Constructors

RawResponse

Fields

rawResponse :: Response

newtype End next

Only accept the end of a path.

Constructors

End

Fields

endNext :: next

newtype WithIO next

Produce a response with IO.

Constructors

WithIO

Fields

withIONext :: IO next

newtype Seg seg next

Match a constant directory in the path.

When specifying APIs, use the /> combinator to specify sub-paths: "foo" /> JSON Bar

Constructors

Seg

Fields

segNext :: next

data OneOfSegs segs next

Match any of a set of path segments.

Constructors

OneOfSegs

Fields

oneOfSegsNext :: next

class FromSegment a where

The class of types that can be parsed from a path segment.

Methods

fromSegment :: Text -> Maybe a

newtype Capture a next

Capture a path segment and pass it on.

Constructors

Capture

Fields

captureNext :: a -> next

newtype Method method next

Accepts requests with a certain method.

Constructors

Method

Fields

methodNext :: next

data ExtraHeaders next

Set extra headers on responses. Existing headers will be overriden if specified here.

Constructors

ExtraHeaders

Fields

extraHeaders :: ResponseHeaders

extraHeadersNext :: next

newtype NoCache next

Prevent caching for sub-routers.

Constructors

NoCache

Fields

noCacheNext :: next

newtype ReqBodyJSON a next

Parse a JSON request body.

Constructors

ReqBodyJSON

Fields

reqBodyJSONNext :: a -> next

type MultiPartData = ([Param], [File FilePath])

A parsed "multipart/form-data" request.

data ReqBodyMultipart a next

Accept a "multipart/form-data" request. Files will be stored in a temporary directory and will be deleted automatically after the request is processed.

Constructors

ReqBodyMultipart

Fields

reqMultiPartParse :: MultiPartData -> Either String a

reqMultiPartNext :: a -> next

type Endpoint method a = End :> (NoCache :> (Method method :> (WithIO :> a)))

Useful synonym for dynamic endpoints: accept requests with a given method, compute a JSON response in IO and don't cache.

data left :<|> right infixr 1

Try to route with left, or try to route with right.

Constructors

(:<|>) infixr 1

Fields

altLeft :: left

altRight :: right

Abbreviation

class Abbreviated a where

Most Routers are really just newtypes. By using brief, you can construct trees of Routers by providing only their inner types, much like Servant.

Minimal complete definition

Nothing

Associated Types

type Brief a :: *

Methods

brief :: Brief a -> a

Error handling

data SolgaError

A Router-related exception with a corresponding HTTP error code.

badRequest :: Text -> SolgaError

Create a 400 Bad Request error with a given message.

notFound :: Text -> SolgaError

Create a 404 Not Found error with a given message.

Router implementation

class Router r where

Routers are the basic typeclass of Solga: their types describe what type of requests they accept, and their values describe how to handle them.

You can use Generic to get free instance of Router for any data type with one constructor and Routers as fields. The fields will be considered alternatives, as if you wrote :<|> between them.

Minimal complete definition

Nothing

Methods

tryRoute :: Request -> Maybe (r -> Responder)

Given a request, if the router supports the given request return a function that constructs a response with a concrete router.

type Responder = (Response -> IO ResponseReceived) -> IO ResponseReceived

The right hand side of Application. Request is already known.

tryRouteNext :: Router r' => (r -> r') -> Request -> Maybe (r -> Responder)

Try to route using a type r by providing a function to turn it into a Router r'. Useful for passing routing on to the next step.

tryRouteNextIO :: Router r' => (r -> IO r') -> Request -> Maybe (r -> Responder)

Like tryRouteNext but in IO.